fix: pre-commit
