+++ aqueduct/gateway/decorators/__init__.py
@@ -1,0 +1,35 @@
+from gateway.decorators.auth import check_limits, token_authenticated, tos_accepted
+from gateway.decorators.availability import check_model_availability
+from gateway.decorators.body import FileSizeError, ensure_usage, parse_body, resolve_alias
+from gateway.decorators.chat_completions import normalize_reasoning_fields, process_file_content
+from gateway.decorators.errors import catch_router_exceptions
+from gateway.decorators.files import process_batch_file, require_files_api_client
+from gateway.decorators.log import log_request
+from gateway.decorators.mcp import (
+    check_mcp_server_availability,
+    mcp_transport_security,
+    parse_jsonrpc_message,
+)
+from gateway.decorators.responses import check_tool_availability, validate_response_id
+
+__all__ = [
+    "FileSizeError",
+    "catch_router_exceptions",
+    "check_limits",
+    "check_mcp_server_availability",
+    "check_model_availability",
+    "check_tool_availability",
+    "ensure_usage",
+    "log_request",
+    "mcp_transport_security",
+    "normalize_reasoning_fields",
+    "parse_body",
+    "parse_jsonrpc_message",
+    "process_batch_file",
+    "process_file_content",
+    "require_files_api_client",
+    "resolve_alias",
+    "token_authenticated",
+    "tos_accepted",
+    "validate_response_id",
+]
