+++ aqueduct/gateway/views/batches.py
@@ -9,15 +9,15 @@ from openai.types.batch_create_params import BatchCreateParams
 from pydantic import TypeAdapter
 
 from gateway.config import get_files_api_client
-from management.models import Batch, BatchStatus, FileObject, Token
-
-from .decorators import (
+from gateway.decorators import (
     catch_router_exceptions,
     log_request,
     parse_body,
     token_authenticated,
     tos_accepted,
 )
+from management.models import Batch, BatchStatus, FileObject, Token
+
 from .errors import error_response
 from .files import sync_batch_file_if_needed
 from .utils import RawJsonResponse
