+++ aqueduct/gateway/views/chat_completions.py
@@ -9,9 +9,7 @@ from litellm.types.utils import ModelResponse
 from pydantic import TypeAdapter
 
 from gateway.config import get_router
-from management.models import Request
-
-from .decorators import (
+from gateway.decorators import (
     catch_router_exceptions,
     check_limits,
     check_model_availability,
@@ -24,6 +22,8 @@ from .decorators import (
     token_authenticated,
     tos_accepted,
 )
+from management.models import Request
+
 from .utils import RawJsonResponse, RawStreamingResponse, get_token_usage
