+++ aqueduct/gateway/views/completions.py
@@ -9,9 +9,7 @@ from litellm.types.utils import TextCompletionResponse
 from pydantic import TypeAdapter
 
 from gateway.config import get_router
-from management.models import Request
-
-from .decorators import (
+from gateway.decorators import (
     catch_router_exceptions,
     check_limits,
     check_model_availability,
@@ -22,6 +20,8 @@ from .decorators import (
     token_authenticated,
     tos_accepted,
 )
+from management.models import Request
+
 from .utils import RawJsonResponse, RawStreamingResponse, get_token_usage
