+++ aqueduct/gateway/views/embeddings.py
@@ -7,9 +7,7 @@ from django.views.decorators.http import require_POST
 from pydantic import TypeAdapter
 
 from gateway.config import get_router
-from management.models import Request
-
-from .decorators import (
+from gateway.decorators import (
     catch_router_exceptions,
     check_limits,
     check_model_availability,
@@ -20,6 +18,8 @@ from .decorators import (
     token_authenticated,
     tos_accepted,
 )
+from management.models import Request
+
 from .utils import RawJsonResponse, get_token_usage
 
 if TYPE_CHECKING:
