+++ aqueduct/gateway/views/models.py
@@ -7,9 +7,9 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_GET
 
 from gateway.config import get_router_config
+from gateway.decorators import log_request, token_authenticated, tos_accepted
 from management.models import Request, Token
 
-from .decorators import log_request, token_authenticated, tos_accepted
 from .utils import RawJsonResponse
 
 MODEL_CREATION_TIMESTAMP = int(timezone.now().timestamp())
