+++ aqueduct/gateway/views/transcriptions.py
@@ -10,9 +10,7 @@ from openai.types.audio.transcription_create_params import (
 )
 from pydantic import ConfigDict, RootModel, TypeAdapter
 
-from management.models import Request
-
-from .decorators import (
+from gateway.decorators import (
     catch_router_exceptions,
     check_limits,
     check_model_availability,
@@ -22,6 +20,8 @@ from .decorators import (
     token_authenticated,
     tos_accepted,
 )
+from management.models import Request
+
 from .utils import RawJsonResponse, RawStreamingResponse, get_token_usage, oai_client_from_body
