+++ aqueduct/gateway/views/vector_store_file_batches.py
@@ -12,6 +12,13 @@ from openai.types.vector_stores.file_batch_create_params import FileBatchCreateP
 from pydantic import TypeAdapter
 
 from gateway.config import get_files_api_client
+from gateway.decorators import (
+    catch_router_exceptions,
+    log_request,
+    parse_body,
+    token_authenticated,
+    tos_accepted,
+)
 from management.models import (
     FileObject,
     Token,
@@ -22,13 +29,6 @@ from management.models import (
     VectorStoreFileStatus,
 )
 
-from .decorators import (
-    catch_router_exceptions,
-    log_request,
-    parse_body,
-    token_authenticated,
-    tos_accepted,
-)
 from .errors import error_response
 from .utils import RawJsonResponse
