+++ aqueduct/gateway/views/vector_stores.py
@@ -15,9 +15,7 @@ from openai.types.vector_store_update_params import VectorStoreUpdateParams
 from pydantic import TypeAdapter
 
 from gateway.config import get_files_api_client
-from management.models import Token, VectorStore
-
-from .decorators import (
+from gateway.decorators import (
     catch_router_exceptions,
     log_request,
     parse_body,
@@ -25,6 +23,8 @@ from .decorators import (
     token_authenticated,
     tos_accepted,
 )
+from management.models import Token, VectorStore
+
 from .errors import error_response
 from .utils import RawJsonResponse
